feat: Make root.go the main entry point file and move shell.go to its folder
